Rename app-sidebar folder to sidebar
