+++ web-app/src/app/(app)/components/sidebar/components/agent-add-button.tsx
