+++ web-app/src/app/(app)/components/sidebar/index.tsx
